backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
supabase==2.3.4
python-dotenv==1.0.0
gotrue==2.9.0
httpx==0.27.0
postgrest==0.16.0
realtime==1.0.0
storage3==0.7.0
